+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/impl/GitRepoManagerImpl.java
@@ -110,6 +110,23 @@ public class GitRepoManagerImpl implements GitRepoManager {
     }
 
     /**
+     * Unpack the git archive of the project.
+     *
+     * @param group   Project group. Must not be {@literal null}.
+     * @param project Project. Must not be {@literal null}.
+     * @throws IOException     If an IO error occurred.
+     * @throws GitAPIException The exception is caused by the internal functionality of managing git repositories.
+     */
+    @Override
+    public void gCheckout(String group, String project) throws IOException, GitAPIException {
+        final Path sourceGitRepository = HelperGitRepo.getSourceGitRepository(group, project);
+
+        Git.open(sourceGitRepository.toFile())
+                .checkout()
+                .call();
+    }
+
+    /**
      * Open and get the contents of a file from a git repository by group and project name.
      *
      * @param group   The name of the git repository group.
@@ -177,7 +194,7 @@ public class GitRepoManagerImpl implements GitRepoManager {
      * @return Search result in git repository.
      */
     @Override
-    public List<ContainerInfoSearchGitRepo> searchByGrep(GitRepoFilterSearch filterSearch) {
+    public List<ContainerInfoSearchGitRepo> searchByGrep_jgit(GitRepoFilterSearch filterSearch) {
         final List<ContainerInfoSearchGitRepo> searchFileGitRepos = new ArrayList<>();
         filterSearch.getSources()
                 .forEach(source -> {
@@ -213,7 +230,7 @@ public class GitRepoManagerImpl implements GitRepoManager {
      * @return Search result in git repository.
      */
     @Override
-    public List<ContainerInfoSearchGitRepo> searchByGrepVirtualThreads(GitRepoFilterSearch filterSearch) {
+    public List<ContainerInfoSearchGitRepo> searchByGrepVirtualThreads_jgit(GitRepoFilterSearch filterSearch) {
         final List<ContainerInfoSearchGitRepo> searchFileGitRepos = new ArrayList<>();
 
         try (final ExecutorService executor = Executors.newVirtualThreadPerTaskExecutor()) {
@@ -262,7 +279,7 @@ public class GitRepoManagerImpl implements GitRepoManager {
             Path source,
             ContainerGitRepoMeta gitRepo) throws IOException, SearchEngineGitSetMaxCountException,
             SearchEngineGitSetMaxDepthException, SearchEngineGitSetContextException {
-        final SearchEngineGrep commandGrep = new SearchEngineGitGrepImpl(
+        final SearchEngineGrep commandGrep = new SearchEngineJGitGrepImpl(
                 filterSearch.getPattern(),
                 source,
                 gitRepo